requirements.txt
nltk
python-Levenshtein
matplotlib
numpy
seaborn
requests
more_itertools
networkx
sklearn
yellowbrick
dataframe-image